Make sure that seeder use id

+++ src/Migrations/2021_03_12_073541_create_email_resets.php
@@ -14,8 +14,8 @@ class CreateEmailResets extends Migration
     public function up()
     {
         Schema::create(config('badaso.database.prefix').'email_resets', function (Blueprint $table) {
-            $table->increments('id');
-            $table->bigInteger('user_id')->nullable();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id')->nullable();
             $table->string('email');
             $table->string('verification_token')->nullable()->unique();
             $table->dateTime('expired_at')->nullable();
